CloudReady shim 15.4 submission
These are the artifacts that are required to submit
CloudReady's build of shim 15.4 for review.  This shim submission is
necessary to fully remediate the BootlHole vulnerability from 2020 and
revokes vulnerable versions of GRUB by using a new EV code signing cert
for the first time.

Shim binaries, build.log with redacted AWS credentials and the required,
fully filled-out README.md are all present.